Consolidated app.py code to cloud_tracking
Transferred all the relevant image processing functions from app.py to cloud_tracking.py
